+++ lib/src/controls/better_player_material_controls.dart
@@ -7,7 +7,7 @@ import 'package:better_player_plus/src/controls/better_player_multiple_gesture_d
 import 'package:better_player_plus/src/controls/better_player_progress_colors.dart';
 import 'package:better_player_plus/src/core/better_player_controller.dart';
 import 'package:better_player_plus/src/core/better_player_utils.dart';
-import 'package:better_player_plus/src/video_player/video_player.dart';
+import 'package:better_player_plus/src/platform_video/video_player.dart';
 
 // Flutter imports:
 import 'package:flutter/material.dart';