+++ lib/src/core/better_player_controller.dart
@@ -5,8 +5,8 @@ import 'package:better_player_plus/src/configuration/better_player_controller_ev
 import 'package:better_player_plus/src/core/better_player_utils.dart';
 import 'package:better_player_plus/src/subtitles/better_player_subtitle.dart';
 import 'package:better_player_plus/src/subtitles/better_player_subtitles_factory.dart';
-import 'package:better_player_plus/src/video_player/video_player.dart';
-import 'package:better_player_plus/src/video_player/video_player_platform_interface.dart';
+import 'package:better_player_plus/src/platform_video/video_player.dart';
+import 'package:better_player_plus/src/platform_video/video_player_platform_interface.dart';
 import 'package:collection/collection.dart' show IterableExtension;
 import 'package:flutter/material.dart';
 import 'package:path_provider/path_provider.dart';
@@ -541,10 +541,7 @@ class BetterPlayerController {
         }
         break;
 
-      default:
-        throw UnimplementedError(
-            "${betterPlayerDataSource.type} is not implemented");
-    }
+      }
     await _initializeVideo();
   }